feat(cache): create CacheModule with Redis and in-memory client fallback
