+++ src/app.module.ts
@@ -8,8 +8,7 @@ import { TasksModule } from './modules/task/tasks.module';
 import { AuthModule } from './modules/auth/auth.module';
 import { RateLimitGuard } from './common/guards/rate-limit.guard';
 import { APP_GUARD } from '@nestjs/core';
-import { InMemoryCacheClient } from './common/cache/in-memory-cache.client';
-import { RedisCacheClient } from './common/cache/redis-cache.client';
+import { CacheModule } from './common/cache/cache.module';
 import { CacheInterceptor } from './common/interceptors/cache.interceptor';
 import { RequestLoggingMiddleware } from './common/middleware/request-logging.middleware';
 import { LoggingModule } from './modules/logging/logging.module';
@@ -36,6 +35,7 @@ import { LogsModule } from './modules/logging/logs.module';
       },
     }),
     TasksModule,
+    CacheModule,
     // register auth module so JwtStrategy and JwtModule are available
     AuthModule,
     LoggingModule,
@@ -45,24 +45,6 @@ import { LogsModule } from './modules/logging/logs.module';
   providers: [
     AppService,
     { provide: APP_GUARD, useClass: RateLimitGuard },
-    {
-      provide: 'CACHE_CLIENT',
-      useFactory: () => {
-        const host = process.env.REDIS_HOST;
-        const port = Number(process.env.REDIS_PORT ?? 6379);
-        if (host) {
-          try {
-            return new RedisCacheClient(host, port);
-          } catch (e) {
-            // Fail open: log if needed and fallback to in-memory client
-            // eslint-disable-next-line no-console
-            console.warn('RedisCacheClient build failed, falling back to in-memory cache', e?.message ?? e);
-            return new InMemoryCacheClient();
-          }
-        }
-        return new InMemoryCacheClient();
-      },
-    },
     CacheInterceptor,
   ],
 })
